Refactore
