+++ api/server.go
@@ -2,8 +2,6 @@ package api
 
 import (
 	"guideventureapi/db"
-	"log"
-	"net/http"
 
 	"github.com/gin-gonic/gin"
 )
@@ -13,67 +11,39 @@ type Server struct {
 	db         db.Database
 }
 
-func NewServer(listenAddr string, createDummyData bool) (*Server, error) {
-	db, err := db.NewSQLiteDb()
-	if err != nil {
-		return nil, err
-	}
+type Option func(*Server) error
 
-	if createDummyData {
-		err = db.CreateDummyData()
-		if err != nil {
-			log.Panic(err)
+func NewServer(options ...Option) (*Server, error) {
+	server := &Server{}
+	for _, opt := range options {
+		if err := opt(server); err != nil {
+			return nil, err
 		}
 	}
-
-	return &Server{
-		listenAddr: listenAddr,
-		db:         db,
-	}, nil
+	return server, nil
 }
 
-func (s *Server) GetGames(c *gin.Context) {
-	games, err := s.db.GetGames()
-	if err != nil {
-		c.AbortWithError(http.StatusInternalServerError, err)
-		return
+func WithListenAddr(listenAddr string) Option {
+	return func(s *Server) error {
+		s.listenAddr = listenAddr
+		return nil
 	}
-
-	c.JSON(http.StatusOK, &games)
 }
 
-func (s *Server) GetGame(c *gin.Context) {
-	gameId := c.Param("gameId")
-	game, err := s.db.GetGame(gameId)
-	if err != nil {
-		c.AbortWithError(http.StatusInternalServerError, err)
-		return
+func WithDatabase(db db.Database, err error) Option{
+	return func(s *Server) error {
+		if err != nil {
+			return err
+		}
+		s.db = db
+		return nil
 	}
-
-	c.JSON(http.StatusOK, &game)
 }
 
-func (s *Server) GetSteps(c *gin.Context) {
-	gameId := c.Param("gameId")
-	steps, err := s.db.GetSteps(gameId)
-	if err != nil {
-		c.AbortWithError(http.StatusInternalServerError, err)
-		return
+func WithDummyData() Option {
+	return func(s *Server) error {
+		return s.db.CreateDummyData()
 	}
-
-	c.JSON(http.StatusOK, &steps)
-}
-
-func (s *Server) GetStep(c *gin.Context) {
-	gameId := c.Param("gameId")
-	stepIndex := c.Param("stepIndex")
-	step, err := s.db.GetStep(gameId, stepIndex)
-	if err != nil {
-		c.AbortWithError(http.StatusInternalServerError, err)
-		return
-	}
-
-	c.JSON(http.StatusOK, &step)
 }
 
 func (s *Server) Start() error {
